[Silver II] Title: 랜선 자르기, Time: 4 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/1654. 랜선 자르기/랜선 자르기.cc
@@ -27,7 +27,7 @@ int main() {
 	while (low <= high) {
 		long long mid = (low + high) / 2;
 
-		long long cnt = Cut(mid);
+		int cnt = Cut(mid);
 		if (cnt >= n) {
 			answer = max(answer, mid);
 			low = mid + 1;
